requirements.txt
boto==2.46.1
Django==1.10.5
django-allauth==0.30.0
django-dotenv==1.4.1
django-light-enums==0.1.6
django-nested-admin==3.0.12
django-ses==0.8.1
django-shortuuidfield==0.1.3
django-s3-storage==0.11.0
django-webpack-loader==0.4.1
django-webtest==1.8.0
djangorestframework==3.5.3
djangorestframework-expander==0.2.3
factory-boy==2.8.1
gunicorn==19.6.0
mysqlclient==1.3.9
Pillow==4.0.0
raven==6.0.0
